Move getBaseAttributes func and test to it's own file/folder
