+++ backend-api/src/functions/common/session/updateOperations/BiometricTokenIssued/BiometricTokenIssued.ts
@@ -1,13 +1,12 @@
 import { UpdateSessionOperation } from "../UpdateSessionOperation";
 import { DocumentType } from "../../../../types/document";
-import { SessionState, BaseSessionAttributes } from "../../session";
+import { SessionState } from "../../session";
 import {
   AttributeValue,
   ReturnValue,
   ReturnValuesOnConditionCheckFailure,
 } from "@aws-sdk/client-dynamodb";
-import { NativeAttributeValue, unmarshall } from "@aws-sdk/util-dynamodb";
-import { emptyFailure, Result, successResult } from "../../../../utils/result";
+import { getBaseSessionAttributes } from "../sessionAttributes/sessionAttributes";
 
 export class BiometricTokenIssued implements UpdateSessionOperation {
   constructor(
@@ -44,36 +43,3 @@ export class BiometricTokenIssued implements UpdateSessionOperation {
     return getBaseSessionAttributes(item);
   }
 }
-
-const getBaseSessionAttributes = (
-  item: Record<string, AttributeValue> | undefined,
-): Result<BaseSessionAttributes, void> => {
-  if (item == null) return emptyFailure();
-
-  const sessionAttributes = unmarshall(item);
-  if (!isBaseSessionAttributes(sessionAttributes)) return emptyFailure();
-
-  return successResult(sessionAttributes);
-};
-
-const isBaseSessionAttributes = (
-  item: Record<string, NativeAttributeValue>,
-): item is BaseSessionAttributes => {
-  if (typeof item.clientId !== "string") return false;
-  if (typeof item.govukSigninJourneyId !== "string") return false;
-  if (typeof item.createdAt !== "number") return false;
-  if (typeof item.issuer !== "string") return false;
-  if (typeof item.sessionId !== "string") return false;
-  if (typeof item.sessionState !== "string") return false;
-  if (typeof item.clientState !== "string") return false;
-  if (typeof item.subjectIdentifier !== "string") return false;
-  if (typeof item.timeToLive !== "number") return false;
-  if (
-    "redirectUri" in item &&
-    item.redirectUri != null &&
-    typeof item.redirectUri !== "string"
-  ) {
-    return false;
-  }
-  return true;
-};
